Fix record cache on client created records
The record cache is updated fine when a record is retrieved from
server, but when a record was created on the client side the record
cache, which is updated in `updateRecord` was simply a reference
local to the `updateRecord` method. Now we use a method to retrieve and
create the adapter stored record cache `_getRecordCache`.